+++ recipe/schemas/parsed_constructors.py
@@ -11,13 +11,13 @@ from .field_grammar import (
     full_condition_parser,
 )
 from .utils import (
-    aggregations,
-    conversions,
+    generate_lookup_by_engine,
     find_column,
     ingredient_class_for_name,
     convert_value,
     calc_date_range,
 )
+from .engine_support import aggregations_by_engine, conversions_by_engine
 from recipe.exceptions import BadIngredient
 from recipe.ingredients import InvalidIngredient
 
@@ -31,6 +31,18 @@ class TransformToSQLAlchemyExpression(Transformer):
     def __init__(self, selectable, require_aggregation=False):
         self.selectable = selectable
         self.require_aggregation = require_aggregation
+        # Database driver
+        try:
+            self.drivername = selectable.metadata.bind.url.drivername
+        except:
+            self.drivername = "unknown"
+
+        self.aggregations = generate_lookup_by_engine(
+            aggregations_by_engine, self.drivername
+        )
+        self.conversions = generate_lookup_by_engine(
+            conversions_by_engine, self.drivername
+        )
 
     def number(self, value):
         try:
@@ -60,7 +72,18 @@ class TransformToSQLAlchemyExpression(Transformer):
         return "BETWEEN"
 
     def aggregate(self, name):
-        return aggregations.get(name.lower())
+        """Return a callable that generates SQLAlchemy to aggregate a field.
+
+        Aggregations may be database specific
+        """
+        ag = self.aggregations.get(name.lower())
+        if ag is None:
+            raise ValueError(
+                "Aggregation {} is not supported on {} engine".format(
+                    name, self.drivername
+                )
+            )
+        return ag
 
     def div(self, num, denom):
         """SQL safe division"""
@@ -89,10 +112,23 @@ class TransformToSQLAlchemyExpression(Transformer):
             return aggr(val)
 
     def conversion(self, name):
-        return conversions.get(name.lower())
+        conv_fn = self.conversions.get(name.lower())
+        if conv_fn is None:
+            raise ValueError(
+                "Conversion {} is not supported on {} engine".format(
+                    conv_fn, self.drivername
+                )
+            )
+        return conv_fn
 
     def convertedcol(self, conversion, col):
-        conv_fn = conversions.get(conversion.lower())
+        conv_fn = self.conversions.get(conversion.lower())
+        if conv_fn is None:
+            raise ValueError(
+                "Conversion {} is not supported on {} engine".format(
+                    conv_fn, self.drivername
+                )
+            )
         return conv_fn(col)
 
     def expr(self, expr):
@@ -114,11 +150,11 @@ class TransformToSQLAlchemyExpression(Transformer):
 
     def is_comparison(self, *args):
         """Things that can be compared with IS
-        
+
         will be either IS NULL or an intelligent date
         like "IS prior year" or "IS next week".
 
-        If intelligent dates are used, determine the relevant dates and 
+        If intelligent dates are used, determine the relevant dates and
         return a tuple of start_date, end_date
         """
         if len(args) == 1 and args[0] is None:
@@ -247,5 +283,10 @@ def create_ingredient_from_parsed(ingr_dict, selectable):
     try:
         return IngredientClass(*args, **ingr_dict)
     except BadIngredient as e:
-        error = {"type": "bad_ingredient", "extra": {"details": str(e),}}
+        error = {
+            "type": "bad_ingredient",
+            "extra": {
+                "details": str(e),
+            },
+        }
         return InvalidIngredient(error=error)