+++ recipe/schemas/utils.py
@@ -1,13 +1,13 @@
+from copy import copy
 from datetime import date, datetime
 from dateutil.relativedelta import relativedelta
 import dateparser
 import inspect
-from sqlalchemy import distinct, func, cast, String, Integer
 from sqlalchemy.ext.declarative import DeclarativeMeta
 from sqlalchemy.sql.base import ImmutableColumnCollection
 from sureberus import schema as S
 
-from recipe.exceptions import BadIngredient, InvalidColumnError
+from recipe.exceptions import InvalidColumnError
 from recipe.compat import basestring
 
 SCALAR_TYPES = [S.Integer(), S.String(), S.Float(), S.Boolean()]
@@ -56,46 +56,29 @@ def coerce_format(v):
     return format_lookup.get(v, v)
 
 
-# Aggregations are a callable on a column expressoin that yields an
-# aggregated column expression
-# for instance sum(sales) => func.sum(MyTable.sales)
-aggregations = {
-    "sum": func.sum,
-    "min": func.min,
-    "max": func.max,
-    "avg": func.avg,
-    "count": func.count,
-    "count_distinct": lambda fld: func.count(distinct(fld)),
-    # Technically "none" is not an aggregation but we're keeping
-    # it here for backward compatibility
-    "none": lambda fld: fld,
-    None: lambda fld: fld,
-    # Percentile aggregations do not work in all engines
-    "median": func.median,
-    "percentile1": lambda fld: func.percentile_cont(0.01).within_group(fld),
-    "percentile5": lambda fld: func.percentile_cont(0.05).within_group(fld),
-    "percentile10": lambda fld: func.percentile_cont(0.10).within_group(fld),
-    "percentile25": lambda fld: func.percentile_cont(0.25).within_group(fld),
-    "percentile50": lambda fld: func.percentile_cont(0.50).within_group(fld),
-    "percentile75": lambda fld: func.percentile_cont(0.75).within_group(fld),
-    "percentile90": lambda fld: func.percentile_cont(0.90).within_group(fld),
-    "percentile95": lambda fld: func.percentile_cont(0.95).within_group(fld),
-    "percentile99": lambda fld: func.percentile_cont(0.99).within_group(fld),
-}
+def generate_lookup_by_engine(lookup_by_engine, engine):
+    """Convert `aggregations_by_engine` or `conversions_by_engine`
+    into a dictionary specific to the provided engine/drivername.
+    """
+    result = copy(lookup_by_engine.get("default", {}))
+    result.update(lookup_by_engine.get(engine, {}))
+    for k, v in list(result.items()):
+        if v is None:
+            result.pop(k, None)
+    return result
 
-# Conversions are a callable on a column expression that yields a
-# nonaggregated column expression
-# for instance, quarter(sales_date) => func.date_trunc('quarter', MyTable.sales_date)
-conversions = {
-    "month": lambda fld: func.date_trunc("month", fld),
-    "week": lambda fld: func.date_trunc("week", fld),
-    "year": lambda fld: func.date_trunc("year", fld),
-    "quarter": lambda fld: func.date_trunc("quarter", fld),
-    "string": lambda fld: func.cast(fld, String()),
-    "int": lambda fld: func.cast(fld, Integer()),
-    # age doesn't work on all databases
-    "age": lambda fld: func.date_part("year", func.age(fld)),
-}
+
+def convert_by_engine_keys_to_regex(lookup_by_engine):
+    """ Convert all the keys in a lookup_by_engine to a regex """
+    keys = set()
+    for d in lookup_by_engine.values():
+        for k in d.keys():
+            if isinstance(k, basestring):
+                keys.add(k)
+
+    keys = list(keys)
+    keys.sort(key=lambda item: (len(item), item), reverse=True)
+    return "|".join(keys)
 
 
 def coerce_pop_version(shelf):
@@ -224,7 +207,7 @@ def date_offset(dt, offset, **offset_params):
 
 def calc_date_range(offset, units, dt):
     """Create an intelligent date range using offsets, units and a starting date
-    
+
     Args:
 
         offset: An offset
@@ -240,7 +223,7 @@ def calc_date_range(offset, units, dt):
             day: The provided date
         dt
             The date that will be used for calculations
-            
+
     Returns:
 
         A tuple of dates constructed using the offsets and units