+++ CHANGELOG.rst
@@ -2,6 +2,13 @@
 Changelog
 =========
 
+0.19.0 (2020-09-04)
+-----------------------------------------
+* Support and documentation for compound selection in automatic filters
+* Support for different sqlalchemy generation when using parsed fields
+* Add support for date conversions and percentiles in bigquery.
+* Ingredient.build_filters now returns SQLAlchemy BinaryExpression rather than Filter objects.
+
 0.18.1 (2020-08-07)
 -----------------------------------------
 * Fix a bug in filter binning
@@ -16,7 +23,7 @@ Changelog
 -----------------------------------------
 * Fix Paginate search to use value roles
 
-  0.17.1 (2020-07-09)
+0.17.1 (2020-07-09)
 -----------------------------------------
 * Fix parsed syntax for `field IS NULL`